Renaming and moving refactorings.
