--- src/main/java/com/iscte/engsoft/grupob/CalendarApp/model/CalendarFormat.java
@@ -1,6 +1,0 @@
-package com.iscte.engsoft.grupob.CalendarApp.model;
-
-public enum CalendarFormat {
-    CSV,
-    JSON
-}
