+++ src/main/java/com/iscte/engsoft/grupob/calendarapp/util/CSVHandler.java
@@ -1,4 +1,4 @@
-package com.iscte.engsoft.grupob.CalendarApp.util;
+package com.iscte.engsoft.grupob.calendarapp.util;
 
 import java.io.*;
