+++ src/main/java/com/iscte/engsoft/grupob/calendarapp/util/UrlReader.java
@@ -1,4 +1,4 @@
-package com.iscte.engsoft.grupob.CalendarApp.utils;
+package com.iscte.engsoft.grupob.calendarapp.util;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Component;
