+++ src/test/java/com/iscte/engsoft/grupob/calendarapp/controller/CalendarControllerTest.java
@@ -1,11 +1,11 @@
-package com.iscte.engsoft.grupob.CalendarApp.controller;
+package com.iscte.engsoft.grupob.calendarapp.controller;
 
 import java.io.IOException;
 
-import com.iscte.engsoft.grupob.CalendarApp.model.CalendarFormat;
-import com.iscte.engsoft.grupob.CalendarApp.model.ConsumeURLCalendarRequest;
-import com.iscte.engsoft.grupob.CalendarApp.model.UploadCalendarFileRequest;
-import com.iscte.engsoft.grupob.CalendarApp.utils.UrlReader;
+import com.iscte.engsoft.grupob.calendarapp.model.CalendarFormat;
+import com.iscte.engsoft.grupob.calendarapp.model.ConsumeURLCalendarRequest;
+import com.iscte.engsoft.grupob.calendarapp.model.UploadCalendarFileRequest;
+import com.iscte.engsoft.grupob.calendarapp.util.UrlReader;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
