+++ src/test/java/com/iscte/engsoft/grupob/calendarapp/util/CSVConverterTest.java
@@ -1,6 +1,5 @@
-package com.iscte.engsoft.grupob.CalendarApp;
+package com.iscte.engsoft.grupob.calendarapp.util;
 
-import com.iscte.engsoft.grupob.CalendarApp.util.CSVConverter;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
