+++ src/test/java/com/iscte/engsoft/grupob/calendarapp/util/JSONConverterTest.java
@@ -1,6 +1,5 @@
-package com.iscte.engsoft.grupob.CalendarApp;
+package com.iscte.engsoft.grupob.calendarapp.util;
 
-import com.iscte.engsoft.grupob.CalendarApp.util.JSONConverter;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.util.ResourceUtils;
